% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryNamespace.R
\name{createQueryNamespace}
\alias{createQueryNamespace}
\title{Create query namespace}
\usage{
createQueryNamespace(
  connectionDetails = NULL,
  connectionHandler = NULL,
  usePooledConnection = FALSE,
  tableSpecification = NULL,
  resultModelSpecificationPath = NULL,
  tablePrefix = "",
  snakeCaseToCamelCase = TRUE,
  ...
)
}
\arguments{
\item{resultModelSpecificationPath}{(optional) csv file or files for tableSpecifications - must conform to table spec
format.}
}
\description{
Create a QueryNamespace instance from either a connection handler or a connectionDetails object
Allows construction with various options not handled by QueryNamespace$new

Note - currently not supported is having multiple table prefixes for multiple table namespaces
}
