% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataModel.R
\name{createResultsDataModel}
\alias{createResultsDataModel}
\title{Create the results data model tables on a database server.}
\usage{
createResultsDataModel(
  connection = NULL,
  connectionDetails = NULL,
  schema,
  sql
)
}
\arguments{
\item{connection}{DatabaseConnector connection instance or null}

\item{connectionDetails}{DatabaseConnector connectionDetails instance or null}

\item{schema}{The schema on the postgres server where the tables will be created.}

\item{sql}{The postgres sql with the results data model DDL.}
}
\description{
Create the results data model tables on a database server.
}
\details{
Only PostgreSQL servers are supported.
}
