% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataModel.R
\name{deleteAllRowsForPrimaryKey}
\alias{deleteAllRowsForPrimaryKey}
\title{Delete results rows for primary key values from database server tables}
\usage{
deleteAllRowsForPrimaryKey(connection, schema, tableName, keyValues)
}
\arguments{
\item{connection}{DatabaseConnector connection instance}

\item{schema}{The schema on the postgres server where the results table exists}

\item{tableName}{Database table name}

\item{keyValues}{Key values of results rows to be deleted}
}
\description{
Delete results rows for primary key values from database server tables
}
\details{
Only PostgreSQL servers are supported.
}
