% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PyFunctions.R
\name{disablePythonUploads}
\alias{disablePythonUploads}
\title{Disable python uploads}
\usage{
disablePythonUploads()
}
\description{
This will stop the use of python in uploaResults - not that this will only work for this R session. If you have set
\code{RMM_USE_PYTHON_UPLOADS} in your .Renviron this will reset the next time you start your R session.
}
