% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PyFunctions.R
\name{install_psycopg2}
\alias{install_psycopg2}
\title{install psycopg2}
\usage{
install_psycopg2(
  envname = Sys.getenv("RMM_PYTHON_ENV", unset = "rmm-uploads"),
  method = "auto",
  ...
)
}
\arguments{
\item{envname}{python virtual environment name. Can be set with system environment variable "RMM_PYTHON_ENV", default is rmm-uploads}
}
\description{
Install psycopg2-binary python package into the specified python virtualenv
}
\seealso{
Other {python}: 
\code{\link{enablePythonUploads}()},
\code{\link{pyPgUploadEnabled}()}
}
\concept{{python}}
