% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PyFunctions.R
\name{pyPgUploadEnabled}
\alias{pyPgUploadEnabled}
\title{are python postgresql uploads enabled?}
\usage{
pyPgUploadEnabled()
}
\description{
are python postgresql uploads enabled?
}
\seealso{
Other {python}: 
\code{\link{enablePythonUploads}()},
\code{\link{install_psycopg2}()}
}
\concept{{python}}
