% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlumberRouter.R
\name{createPlumberRouter}
\alias{createPlumberRouter}
\title{Create a router for a plumber API}
\usage{
createPlumberRouter(tableSpecification, ...)
}
\arguments{
\item{...}{@seealso createQueryNamespace}
}
\description{
create a plumber router for each table within a given table space
}
