% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryNamespace.R
\name{createQueryNamespace}
\alias{createQueryNamespace}
\title{Create query namespace}
\usage{
createQueryNamespace(
  connectionDetails = NULL,
  connectionHandler = NULL,
  usePooledConnection = FALSE,
  tableSpecification = NULL,
  resultModelSpecificationPath = NULL,
  tablePrefix = "",
  snakeCaseToCamelCase = TRUE,
  ...
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{connectionHandler}{ResultModelManager ConnectionHandler or PooledConnectionHandler instance}

\item{usePooledConnection}{Use Pooled database connection instead of standard DatabaseConnector single
connection.}

\item{tableSpecification}{Table specfication data.frame}

\item{resultModelSpecificationPath}{(optional) csv file or files for tableSpecifications - must conform to table spec
format.}

\item{tablePrefix}{String to prefix table names with - default is empty string}

\item{snakeCaseToCamelCase}{convert snakecase results to camelCase field names (TRUE by default)}

\item{...}{Elipsis - use for any additional string keys to replace}
}
\description{
Create a QueryNamespace instance from either a connection handler or a connectionDetails object
Allows construction with various options not handled by QueryNamespace$new

Note - currently not supported is having multiple table prefixes for multiple table namespaces
}
