% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PyFunctions.R
\name{enablePythonUploads}
\alias{enablePythonUploads}
\title{Enable Python Postgres Uploads}
\usage{
enablePythonUploads(...)
}
\arguments{
\item{...}{parameters to pass to py_install}
}
\description{
Step by step install to enable python uploads
}
