% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchemaGenerator.R
\name{generateSqlSchema}
\alias{generateSqlSchema}
\title{Schema generator}
\usage{
generateSqlSchema(csvFilpath, sqlOutputPath, overwrite = FALSE)
}
\arguments{
\item{sqlOutputPath}{File to write sql to.}

\item{overwrite}{Boolean - overwrite existing file?}

\item{csvFilepath}{Path to schema file.}
}
\description{
Take a csv schema definition and create a basic sql script with it.
}
