% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchemaGenerator.R
\name{generateSqlSchema}
\alias{generateSqlSchema}
\title{Schema generator}
\usage{
generateSqlSchema(csvFilpath, sqlOutputPath = NULL, overwrite = FALSE)
}
\arguments{
\item{sqlOutputPath}{File to write sql to.}

\item{overwrite}{Boolean - overwrite existing file?}

\item{csvFilepath}{Path to schema file. Csv file must have the columns:
"table_name", "colum_name", "data_type", "is_required", "primary_key"
Note -}
}
\value{
string containing the sql for the table
}
\description{
Take a csv schema definition and create a basic sql script with it.
}
