% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchemaGenerator.R
\name{generateSqlSchema}
\alias{generateSqlSchema}
\title{Schema generator}
\usage{
generateSqlSchema(
  csvFilepath = NULL,
  schemaDefinition = NULL,
  sqlOutputPath = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{csvFilepath}{Path to schema file. Csv file must have the columns:
"table_name", "column_name", "data_type", "primary_key"}

\item{schemaDefinition}{A schemaDefintiion data.frame` with the columns:
tableName, columnName, dataType, isRequired, primaryKey}

\item{sqlOutputPath}{File to write sql to.}

\item{overwrite}{Boolean - overwrite existing file?}
}
\description{
Take a csv schema definition and create a basic sql script with it.
returns string containing the sql for the table
}
