% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataModel.R
\name{loadResultsDataModelSpecifications}
\alias{loadResultsDataModelSpecifications}
\title{Get specifications from a given file path}
\usage{
loadResultsDataModelSpecifications(filePath)
}
\arguments{
\item{filePath}{path to a valid csv file}
}
\value{
A tibble data frame object with specifications
}
\description{
Get specifications from a given file path
}
