% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionHandler.R
\name{PooledConnectionHandler}
\alias{PooledConnectionHandler}
\title{Pooled Connection Handler}
\description{
Transparently works the same way as a standard connection handler but stores pooled connections.
Useful for long running applications that serve multiple concurrent requests.
}
\section{Super class}{
\code{\link[ResultModelManager:ConnectionHandler]{ResultModelManager::ConnectionHandler}} -> \code{PooledConnectionHandler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PooledConnectionHandler-initConnection}{\code{PooledConnectionHandler$initConnection()}}
\item \href{#method-PooledConnectionHandler-closeConnection}{\code{PooledConnectionHandler$closeConnection()}}
\item \href{#method-PooledConnectionHandler-queryFunction}{\code{PooledConnectionHandler$queryFunction()}}
\item \href{#method-PooledConnectionHandler-clone}{\code{PooledConnectionHandler$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ResultModelManager" data-topic="ConnectionHandler" data-id="dbIsValid"><a href='../../ResultModelManager/html/ConnectionHandler.html#method-ConnectionHandler-dbIsValid'><code>ResultModelManager::ConnectionHandler$dbIsValid()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ResultModelManager" data-topic="ConnectionHandler" data-id="executeFunction"><a href='../../ResultModelManager/html/ConnectionHandler.html#method-ConnectionHandler-executeFunction'><code>ResultModelManager::ConnectionHandler$executeFunction()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ResultModelManager" data-topic="ConnectionHandler" data-id="executeSql"><a href='../../ResultModelManager/html/ConnectionHandler.html#method-ConnectionHandler-executeSql'><code>ResultModelManager::ConnectionHandler$executeSql()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ResultModelManager" data-topic="ConnectionHandler" data-id="finalize"><a href='../../ResultModelManager/html/ConnectionHandler.html#method-ConnectionHandler-finalize'><code>ResultModelManager::ConnectionHandler$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ResultModelManager" data-topic="ConnectionHandler" data-id="getConnection"><a href='../../ResultModelManager/html/ConnectionHandler.html#method-ConnectionHandler-getConnection'><code>ResultModelManager::ConnectionHandler$getConnection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ResultModelManager" data-topic="ConnectionHandler" data-id="initialize"><a href='../../ResultModelManager/html/ConnectionHandler.html#method-ConnectionHandler-initialize'><code>ResultModelManager::ConnectionHandler$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ResultModelManager" data-topic="ConnectionHandler" data-id="queryDb"><a href='../../ResultModelManager/html/ConnectionHandler.html#method-ConnectionHandler-queryDb'><code>ResultModelManager::ConnectionHandler$queryDb()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ResultModelManager" data-topic="ConnectionHandler" data-id="renderTranslateSql"><a href='../../ResultModelManager/html/ConnectionHandler.html#method-ConnectionHandler-renderTranslateSql'><code>ResultModelManager::ConnectionHandler$renderTranslateSql()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PooledConnectionHandler-initConnection"></a>}}
\if{latex}{\out{\hypertarget{method-PooledConnectionHandler-initConnection}{}}}
\subsection{Method \code{initConnection()}}{
Overrides ConnectionHandler Call
Close Connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PooledConnectionHandler$initConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PooledConnectionHandler-closeConnection"></a>}}
\if{latex}{\out{\hypertarget{method-PooledConnectionHandler-closeConnection}{}}}
\subsection{Method \code{closeConnection()}}{
Overrides ConnectionHandler Call
query Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PooledConnectionHandler$closeConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PooledConnectionHandler-queryFunction"></a>}}
\if{latex}{\out{\hypertarget{method-PooledConnectionHandler-queryFunction}{}}}
\subsection{Method \code{queryFunction()}}{
Overrides ConnectionHandler Call. Does not translate or render sql.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PooledConnectionHandler$queryFunction(sql, snakeCaseToCamelCase = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sql}}{sql query string}

\item{\code{snakeCaseToCamelCase}}{(Optional) Boolean. return the results columns in camel case (default)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PooledConnectionHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PooledConnectionHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PooledConnectionHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
