% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultModelManager.R
\docType{package}
\name{ResultModelManager-package}
\alias{ResultModelManager}
\alias{ResultModelManager-package}
\title{ResultModelManager: Result Model Manager (RMM) for OHDSI packages}
\description{
Database data model management utilities for OHDSI packages.
}
\author{
\strong{Maintainer}: James Gilbert \email{gilbert@ohdsi.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
