% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultModelManager.R
\docType{package}
\name{ResultModelManager-package}
\alias{ResultModelManager}
\alias{ResultModelManager-package}
\title{ResultModelManager: Result Model Manager}
\description{
Database data model management utilities for R packages in the Observational Health Data Sciences and Informatics program.
}
\author{
\strong{Maintainer}: Jamie Gilbert \email{gilbert@ohdsi.org}

}
\keyword{internal}
