% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultModelManager.R
\docType{package}
\name{ResultModelManager-package}
\alias{ResultModelManager}
\alias{ResultModelManager-package}
\title{ResultModelManager: Result Model Manager}
\description{
Database data model management utilities for R packages in the Observational Health Data Sciences and Informatics program \url{https://ohdsi.org}. 'ResultModelManager' provides utility functions to allow package maintainers to migrate existing SQL database models, export and import results in consistent patterns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/ResultModelManager}
  \item Report bugs at \url{https://github.com/OHDSI/ResultModelManager/issues}
}

}
\author{
\strong{Maintainer}: Jamie Gilbert \email{gilbert@ohdsi.org}

}
\keyword{internal}
